Introduce generate:deduplicate script
This script removes duplicate entries that would never "win" the
race the matcher runs to find the first match in a given language.
For example, the word "the" will match the en/wikipedia file first
because it's the 1st entry. It's the 2nd entry in en/commonWords,
but that ends up just taking space and increasing the commonWords
list unecessarily.

This script will load all the words for each language, filter out
the duplicate words that have a worse "rank", and then re-write the
files. It runs in ~333econds on my machine. :)