feat(languages): Add deduplication functionality (#54)
* Add support for blank passwords

* Linting fixes - oops!

* Re-run generate:languageData

* Introduce generate:deduplicate script

This script removes duplicate entries that would never "win" the
race the matcher runs to find the first match in a given language.
For example, the word "the" will match the en/wikipedia file first
because it's the 1st entry. It's the 2nd entry in en/commonWords,
but that ends up just taking space and increasing the commonWords
list unecessarily.

This script will load all the words for each language, filter out
the duplicate words that have a worse "rank", and then re-write the
files. It runs in ~333econds on my machine. :)

* Run the deduplicate script

Results of running du -k for packages below:

common: 496 -> 496 (±0%)
de: 1060 -> 932 (-12.1%)
fr: 868 -> 672 (-22.5%)
en: 1496 -> 1336 (-10.7%)
nl-be: 604 -> 480 (-20.5%)
total: 4616 -> 4008 (-13.2%)

* Fix failing test

The rank changed to be lower due to duplicate entries prior to 'do'
being removed from commonWords.